Electrical Wire & Airlock Added
Added texture for wire. Autotile is mostly complete, experiences issues when placing four 4-ways neighboring, causing them to revert to the first tile. Place on TestMap for demostration and later testing.

Added texture for airlock. Airlock inherit is currently in development: AnimationPlayer is complete, however animationTree and script is currently not complete and needs further coding.